cpu v9.0.0-rc.0

+++ src/cpu/src/op/silencer.h
@@ -11,9 +11,9 @@ typedef ALIGN2 struct {
 } ConfigSilencer;
 
 inline static bool_t validate_silencer_settings(
-    const bool_t strict_mode, const uint32_t min_freq_div_intensity,
-    const uint32_t min_freq_div_phase, const uint32_t stm_freq_div,
-    const uint32_t mod_freq_div) {
+    const bool_t strict_mode, const uint16_t min_freq_div_intensity,
+    const uint16_t min_freq_div_phase, const uint16_t stm_freq_div,
+    const uint16_t mod_freq_div) {
   if (strict_mode) {
     if ((mod_freq_div < min_freq_div_intensity) ||
         (stm_freq_div < min_freq_div_intensity) ||
